Start generalizing decimal formatting
We'd like to improve formatting of other types to support decimals. To
avoid code duplication, this commit does part of the change: simple move
to a separate file with appropriate plumbing.

This is still not fully general but it's simpler to review.
